Move transaction related functions from deploy to transact
Split functions that send and wait for a transaction to a function
that sends the transaction and does not wait and one that waits.
